+++ src/components/Sparkline/Sparkline.test.tsx
@@ -15,14 +15,6 @@ describe('<Sparkline />', () => {
         return {matches: false};
       }),
     });
-
-    Object.defineProperty(window, 'ResizeObserver', {
-      value: jest.fn(() => ({
-        observe: jest.fn(),
-        unobserve: jest.fn(),
-        disconnect: jest.fn(),
-      })),
-    });
   });
 
   const mockSeries = [